feat: training first mnist network
Added code for experimenting with tensors and numpy arrays and for training first pyTorch network
Realized that the approach in pytorch_base_network is deemed to fail as pyTorch does not capture the modules defined within a list. 
ModuleList was added to pytorch for this purpose, pytorch_base_network must be modified to use it.